+++ lib/src/custom/main_dialog.dart
@@ -2,8 +2,10 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_smart_dialog/src/data/base_controller.dart';
+import 'package:flutter_smart_dialog/src/data/smart_tag.dart';
 import 'package:flutter_smart_dialog/src/helper/config.dart';
 import 'package:flutter_smart_dialog/src/helper/dialog_proxy.dart';
+import 'package:flutter_smart_dialog/src/helper/route_record.dart';
 import 'package:flutter_smart_dialog/src/widget/attach_dialog_widget.dart';
 import 'package:flutter_smart_dialog/src/widget/smart_dialog_widget.dart';
 
@@ -115,9 +117,10 @@ class MainDialog {
       var tempWidget = _widget;
       showDialog(
         context: DialogProxy.navigatorContext!,
-        barrierColor: Colors.red,
+        barrierColor: Colors.transparent,
         barrierDismissible: false,
         useSafeArea: false,
+        routeSettings: RouteSettings(name: SmartTag.systemDialog),
         builder: (BuildContext context) {
           return tempWidget;
         },