+++ lib/src/data/smart_tag.dart
@@ -1,0 +1,7 @@
+class SmartTag {
+  /// DialogRoute setting name
+  static const String systemDialog = 'smartSystemDialog';
+
+  ///SmartDialog tag：keepSingle
+  static const String keepSingle = 'smartKeepSingle';
+}